pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[project]
name = "econagents-ibex-tudelft"
license = "MIT"
readme = "README.md"
homepage = "https://github.com/IBEX-TUDelft/econagents"
repository = "https://github.com/IBEX-TUDelft/econagents"
dynamic = ["version"]
requires-python = ">=3.10,<4"
dependencies = [
    "econagents[all]>=0.0.0",
]

[tool.poetry]
name = "econagents-ibex-tudelft"
version = "0.0.0"  # Do not change, let poetry-dynamic-versioning handle it.
packages = [{include = "econagents_ibex_tudelft"}]
include = ["econagents_ibex_tudelft/*.so", "econagents_ibex_tudelft/*.pyd"]  # Compiled extensions
license = "MIT"
description = ""
authors = ["Dylan Castillo <dylan@iwanalabs.com>"]

[tool.poetry.build]
generate-setup-file = false

[tool.poetry.group.dev.dependencies]
coverage = {extras = ["toml"], version = ">=5.1"}
pre_commit = ">=2.16.0"
pytest = ">=7.1.2"
pytest-cov = ">=3.0.0"
pytest-mock = ">=3.7.0"
python-dotenv = "^1.1.1"
nest-asyncio = "^1.6.0"
ruff = "^0.12.7"
types-requests = "^2.32.4.20250611"
pytest-asyncio = "^1.1.0"
types-pyyaml = "^6.0.12.20250516"
jupyter = "^1.0.0"

[tool.poetry.group.debug]
optional = true

[tool.poetry.group.debug.dependencies]
ipdb = ">=0.13.9"
line_profiler = ">=3.5.1"

[tool.coverage.run]
branch = true
omit = [
    "tests/*",
    "examples/*",
]

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self.debug:",
    "if debug:",
    "if DEBUG:",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if False:",
    "if __name__ == .__main__.:",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["tests/"]

[tool.ruff]
target-version = 'py39'

exclude = [
    "migrations",
    "__pycache__",
    "manage.py",
    "settings.py",
    "env",
    ".env",
    "venv",
    ".venv",
]

line-length = 120

[tool.ruff.lint]
ignore = [
    "B905",  # zip strict=True; remove once python <3.10 support is dropped.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D401",
    "E402",
    "E501",
    "F401",
    "F811",  # Redefinition of unused name
    "D205",
    "S311",
]


[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "S101",  # use of "assert"
    "S102",  # use of "exec"
    "S106",  # possible hardcoded password.
    "PGH001",  # use of "eval"
]

[tool.codespell]
skip = 'poetry.lock,'
